New exfiltration methods for SOP test and bug fixes
simple.py:
- Access-Control-Allow-Methods and Access-Control-Allow-Headers are not
  enforced and can be overridden from the cdmline.
- Renamed %%ECHO%% to {ECHO}
- Fixed: the origin URL parameter is URL-decoded
- Fixed: empty string for --secrets was causing an IndexError

SOP test:
- removed getData.html: all requests done from within getSecret.html
- added fetching using iframe and canvas (2d and bitmap)
- testing canvas with an image, as text can't render in a canvas
- js/allfuncs.js:
  - getData fetches the data using all CORS combinations and
    exfiltration methods (as opposed to given CORS combo and all request
    methods); a filtered list of exfiltration methods can be provided
  - setting up the requestsLeft monitor is done by getData, not at
    window load
- added test_exfiltrated_data.sh

Bug fixes:
- demos/sop/getSecret.html:
  - URI-encoding and decoding reqOrigin
- js/allfuncs.js:
  - using webtoolkit.info's base64 implementation, which works on all
    tested browsers and supports unicode characters
  - setting DEBUG during script load
  - registerWinOnLoad: binding this during function call (though it
    should be window either way)
  - registerOnLoad: attachEvent not working for script tags in IE, now
    using .onload and .onreadystatechange
  - sendData: handling missing ctype
  - IE8 was firing the onerror handler twice, fixed
  - IE9 was not firing the onload handler for objects (readyState was 4,
    not loaded), fixed
  - because location.replace was not happening immediately, some code
    that was executed from before the redirection to login was messing
    up the counting of requestsLeft (probably related to the messed
    behaviour for onload/onerror handlers, which required clearing of
    the handler once it's called); returning immediately after
    location.replace fixed it
  - display:none was causing objects not to be loaded

Minor/internal additions:
- js/allfuncs.js:
  - added registerOnError
  - sendData can skip base64-encoding if already encoded
  - optional parameters to functions get a default value if omitted or
    given as null; if given as '' they indicate disabled functionality
    (e.g. appendTo in addElement)
  - added more identifying parameters to requests to help with parsing
    of the log

+++ demos/sop/parse_request_log.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-# to debug: uncomment the last line and do
-# egrep 'DEBUG: ([A-Z]+ \/secret|POST \/demos|END: requested.*1)' $TMPFILE
 
 REQFILE="${1:-logs/requests.log}"
 OUTFILE="${2:-logs/requests_result.md}"
@@ -10,124 +8,150 @@ if [[ $(uname) == Darwin* ]] ; then
 else
   AWK='awk'
 fi
+DEBUG=0
 
-${AWK} '
+${AWK} -v debug=$DEBUG '
 BEGIN {
   IGNORECASE=1
-  debug=0
-  printf "| %-20s | %-20s | %-20s | %-20s | %-20s | %-20s | %-20s |\n", "BROWSER", "METHOD", "ORIGIN", "CREDENTIALS", "PREFLIGHT", "COOKIE", "READ BY JS"
-  print "| :------------------: | :------------------: | :------------------: | :------------------: | :------------------: | :------------------: | :------------------: |"
+  printf "| %-20s | %-22s | %-11s | %-11s | %-15s | %-11s | %-11s |\n", "BROWSER", "METHOD", "ORIGIN", "CREDENTIALS", "PREFLIGHT", "COOKIE", "READ BY JS"
+  print "| :------------------: | :--------------------: | :---------: | :---------: | :-------------: | :---------: | :---------: |"
 }
 /[^ ]/ {
   # look for first non-blank line after start
   if (newReq) {
     newReq=0 
-    if (debug) { print "DEBUG: " $0 }
+    if (debug > 1) { print "DEBUG: " $0 }
     requested=match($0,
-      /^(GET|POST|OPTIONS|HEAD|DELETE|PUT) \/secret\/secret\.txt\?origin=(\*|%%ECHO%%)?&creds=([01])&via=([^&]*)[^ ]* HTTP\/[0-9\.]+$/,
+      /^(GET|POST|OPTIONS|HEAD|DELETE|PUT) \/secret\/secret\.[a-z]+\?origin=([^&]*)&creds=([01])&via=([^&]*)&reqBy=([^&]+)[^ ]* HTTP\/[0-9\.]+$/,
       mArr)
     if (requested) {
       method=mArr[1]
-      origin=mArr[2]
+      if (mArr[2] == mArr[5]) {
+        origin="{ECHO}"
+      } else {
+        origin=gensub(/%3a/, ":", "g", tolower(mArr[2]))
+        origin=gensub(/%2f/, "/", "g", origin)
+      }
+      if (debug > 1) { print "DEBUG: origin: " origin }
       creds=mArr[3]
       exfMethod=mArr[4]
     } else {
       exfiltrated=match($0,
-        /^POST \/demos\/sop\/getData.html\?reqURL=[^&]+%2Fsecret%2Fsecret.txt%3Forigin%3D(\*|%25%25ECHO%25%25)?%26creds%3D([01])[^ ]*&post=([01])&via=([^&]*) HTTP\/[0-9\.]+$/,
+        /^POST \/demos\/sop\/getSecret.html\?.*&allowOrigin=([^&]*)&allowCreds=([01])&method=([^&]+)&via=([^&]+) HTTP\/[0-9\.]+$/,
         mArr)
       if (exfiltrated) {
-        origin=gensub(/%25/, "%", "g", mArr[1])
+        origin=gensub(/%3a/, ":", "g", tolower(mArr[1]))
+        origin=gensub(/%2f/, "/", "g", origin)
+        if (debug > 1) { print "DEBUG: origin: " origin }
         creds=mArr[2]
-        if (mArr[3] == 1) { method="POST" }
-        else { method="GET" }
+        method=mArr[3]
         exfMethod=mArr[4]
       }
     }
     if (debug > 1) { print "DEBUG: requested: " requested ", exfiltrated: " exfiltrated }
   }
 }
-{
-  if (requested || exfiltrated) {
-    # look for the Access-Control-Request-Method, or the Cookie
-    # and User-Agent after a request for /secret/secret.txt
-    if (requested && match($0, /^Access-Control-Request-Method: *(.*)/, mArr)) {
-      method=mArr[1]
-      preflight="Y"
+(requested || exfiltrated) {
+  # look for the Access-Control-Request-Method, or the Cookie
+  # after a request for the secret
+  if (requested && match($0, /^Access-Control-Request-Method: *(.*)/, mArr)) {
+    method=mArr[1]
+    if (debug > 1) { print "DEBUG: preflight for " method }
+    preflight="Y"
+  }
+  if (requested && match($0, /^Cookie: *SESSION=/)) {
+    if (debug > 1) { print "DEBUG: Cookie: " $0 }
+    cookie="Y"
+  }
+  # look for the Host after exfiltration and compare to origin (which is allowed)
+  if (exfiltrated && match($0, /^Host: *(.*)/, mArr)) {
+    if (mArr[1] == gensub(/https?:\/\//, "", "1", origin)) {
+      origin="{ECHO}" # clearly shows when the origin is allowed in the table
     }
-    if (requested && match($0, /^Cookie: *SESSION=/)) {
-      if (debug > 1) { print "DEBUG: Cookie: " $0 }
-      cookie="Y"
+  }
+  # look for the User-Agent
+  if (match($0, /^User-Agent: *([^\(]*) *(\([^\)]*\))? *(.*)/, mArr)) {
+    if (mArr[2]) {
+      os=gensub(/^\(|\)$/, "", "g", mArr[2])
+      uaFull=mArr[1] " " mArr[3]
     }
-    if (match($0, /^User-Agent: *([^\(]*) *(\([^\)]*\))? *(.*)/, mArr)) {
-      if (mArr[2]) {
-        os=gensub(/^\(|\)$/, "", "g", mArr[2])
-        uaFull=mArr[1] " " mArr[3]
-      }
-      else {
-        os=""
-        uaFull=mArr[1]
-      }
+    else {
+      os=""
+      uaFull=mArr[1]
+    }
 
-      if (uaFull == "contype") {
-        # IE being an idiot
-        ua=uaOLD
+    if (uaFull == "contype") {
+      # IE being an idiot
+      ua=uaOLD
+    }
+    else {
+      numF=patsplit(uaFull, uaFields, /[A-Za-z]+\/[0-9\.]+/)
+      # split(os, osFields, /; */)
+      if (debug > 1) {
+        printf "DEBUG: UA: os=" os ", numF=" numF ", uaFields=["
+        for (i = 0; i <= numF ; i++) { printf uaFields[i] ", " }
+        print "]"
       }
-      else {
-        numF=patsplit(uaFull, uaFields, /[A-Za-z]+\/[0-9\.]+/)
-        # split(os, osFields, /; */)
-        if (debug > 1) {
-          printf "DEBUG: UA: os=" os ", numF=" numF ", uaFields=["
-          for (i = 0; i <= numF ; i++) { printf uaFields[i] ", " }
-          print "]"
+      ua=""
+      if (numF > 1) {
+        if ( ( uaFields[1] ~ /Mozilla\// && (uaFields[numF] ~ /OPR\// || uaFields[numF-1] == "Opera") ) ||
+          ( uaFields[1] ~ /Opera\// && uaFields[numF] ~ /Version\// ) ) {
+          # OPERA v1,2,3
+          if (debug > 1) { print "DEBUG: UA is Opera v1,2 or 3" }
+          ua="Opera " gensub(/.*\//, "", "1", uaFields[numF])
         }
-        ua=""
-        if (numF > 2) {
-          if (uaFields[1] ~ /Opera\// || uaFields[1] ~ /Mozilla\// && uaFields[numF] ~ /OPR\//) {
-            # OPERA
-            ua="Opera " gensub(/.*\//, "", "1", uaFields[numF])
-          }
-          else if (uaFields[numF-1] ~ /Chrome\//) {
-            # CHROME
-            ua=gensub(/\//, " ", "1", uaFields[numF-1])
-          }
-          else if (uaFields[numF] ~ /(Firefox)\//) {
-            # FIREFOX
-            ua=gensub(/\//, " ", "1", uaFields[numF])
-          }
-          else if (uaFields[numF] ~ /(Edge)\//) {
-            # EDGE
-            ua=gensub(/\//, " ", "1", uaFields[numF])
-          }
-          else if (uaFields[numF] ~ /(Safari)\//) {
-            # SAFARI
-            version=""
-            for (i in uaFields) {
-              if (uaFields[i] ~ /Version\// ) {
-                version=gensub(/.*\//, "", "1", uaFields[i])
-              }
-            }
-            if (! version) {
-              version=gensub(/.*\//, "", "1", uaFields[numF]) # WebKit version
-            }
-            ua=gensub(/\//, " ", "1", uaFields[numF])
-          }
+        else if (uaFields[1] ~ /Opera\//) {
+          # OPERA v4
+          if (debug > 1) { print "DEBUG: UA is Opera v4" }
+          ua=gensub(/\//, " ", "1", uaFields[1])
         }
-        if (! ua) {
-          if (os ~ /MSIE|Trident/) {
-            # INTERNET EXPLORER
-            match(os, /(Windows NT *)([0-9\.]+)/, mArr)
-            WinVersion=mArr[2]
-            match(os, /(MSIE *|rv *: *)([0-9\.]+)/, mArr)
-            IEVersion=mArr[2]
-            if (! IEVersion) {
-              IEVersion="(Unknown)"
+        else if (uaFields[numF-1] ~ /Chrome\//) {
+          # CHROME
+          if (debug > 1) { print "DEBUG: UA is Chrome" }
+          ua=gensub(/\//, " ", "1", uaFields[numF-1])
+        }
+        else if (uaFields[numF] ~ /(Firefox)\//) {
+          # FIREFOX
+          if (debug > 1) { print "DEBUG: UA is Firefox" }
+          ua=gensub(/\//, " ", "1", uaFields[numF])
+        }
+        else if (uaFields[numF] ~ /(Edge)\//) {
+          # EDGE
+          if (debug > 1) { print "DEBUG: UA is Edge" }
+          ua=gensub(/\//, " ", "1", uaFields[numF])
+        }
+        else if (uaFields[numF] ~ /(Safari)\//) {
+          # SAFARI
+          if (debug > 1) { print "DEBUG: UA is Safari" }
+          version=""
+          for (i in uaFields) {
+            if (uaFields[i] ~ /Version\// ) {
+              version=gensub(/.*\//, "", "1", uaFields[i])
             }
-            ua="IE " IEVersion " (Win " WinVersion ")"
           }
-          else {
-            # UNKNOWN
-            ua=uaFull
+          if (! version) {
+            version=gensub(/.*\//, "", "1", uaFields[numF]) # WebKit version
           }
+          ua=gensub(/\//, " ", "1", uaFields[numF])
+        }
+      }
+      if (! ua) {
+        if (os ~ /MSIE|Trident/) {
+          # INTERNET EXPLORER
+          if (debug > 1) { print "DEBUG: UA is IE" }
+          match(os, /(Windows NT *)([0-9\.]+)/, mArr)
+          WinVersion=mArr[2]
+          match(os, /(MSIE *|rv *: *)([0-9\.]+)/, mArr)
+          IEVersion=mArr[2]
+          if (! IEVersion) {
+            IEVersion="(Unknown)"
+          }
+          ua="IE " IEVersion " (Win " WinVersion ")"
+        }
+        else {
+          # UNKNOWN
+          if (debug > 1) { print "DEBUG: UA is Unknown" }
+          ua=uaFull
         }
       }
     }
@@ -163,12 +187,12 @@ BEGIN {
     if (debug) {
       print "DEBUG: END: id: " id ", requested: " requested ", exfiltrated: " exfiltrated
       if (result[id]["ua"]) {
-        printf "DEBUG: END: result[id]=["
-        printf ua ": "
+        print "DEBUG: END: result[id]=["
+        print "DEBUG:   " ua ": "
         for (f in result[id]) {
-          printf f ": " result[id][f] "; "
+          print "DEBUG:     " f ": " result[id][f] "; "
         }
-        print "]"
+        print "DEBUG:  ]"
       }
     }
   }
@@ -178,31 +202,49 @@ BEGIN {
 }
 END {
   for (id in result) {
-    printf "| %-20s | %-20s | %-20s | %-20s | %-20s | %-20s | %-20s |\n", result[id]["ua"], result[id]["method"], result[id]["origin"], result[id]["creds"], result[id]["preflight"], result[id]["cookie"], result[id]["read"]
+    printf "| %-20s | %-22s | %-11s | %-11s | %-15s | %-11s | %-11s |\n", result[id]["ua"], result[id]["method"], result[id]["origin"], result[id]["creds"], result[id]["preflight"], result[id]["cookie"], result[id]["read"]
   }
 }
 ' "${REQFILE}" > "${TMPFILE}"
+[[ $DEBUG -eq 0 ]] || exit 0
+
+############################################################
+DEBUG=0
 
 # Sort the table
 IFS=$'\n' read -d '' -a browsers < <(tail -n+3 "${TMPFILE}" | cut -d\| -f2 | sort -u)
-# IFS=$'\n' read -d '' -a methods < <(tail -n+3 "${TMPFILE}" | cut -d\| -f3 | sort -u)
+# don't sort them by HTTP method, only exfiltration method; keep the order in which requests were sent
+IFS=$'\n' read -d '' -a exfMethods < <(tail -n+3 "${TMPFILE}" | egrep -o 'via [^ )]+' | sort -u)
+IFS=$'\n' read -d '' -a origins < <(tail -n+3 "${TMPFILE}" | cut -d\| -f4 | sort -u)
 sep=$(sed -n '2p' "${TMPFILE}" | tr ':-' '~')
+SPECIAL_CHARS=('\\' '\(' '\)' '\[' '\]' '\.' '\^' '\$' '\+' '\?' '\*' '\|' '\{' '\}')
+
+if [[ $DEBUG -ne 0 ]] ; then
+  echo "browsers:"
+  printf "  %s\n" "${browsers[@]}"
+  echo "exfMethods:"
+  printf "  %s\n" "${exfMethods[@]}"
+  echo "origins:"
+  printf "  %s\n" "${origins[@]}"
+fi
 
 head -n2 "${TMPFILE}" > "${OUTFILE}"
 for browser in "${browsers[@]}" ; do
-  for c in '(' ')' '[' ']' '.' '^' '$' '+' '\?' '\*' '|' '\' '{' '}' ; do
-    browser="${browser//${c}/\\${c}}"
+  for c in "${SPECIAL_CHARS[@]}" ; do
+    browser="${browser//${c}/${c}}"
   done
-  for origin in '\*' '%%ECHO%%' '' ; do
+  for origin in "${origins[@]}" ; do
+    for c in "${SPECIAL_CHARS[@]}" ; do
+      origin="${origin//${c}/${c}}"
+    done
     for creds in 1 0 ; do
-      # don't sort them by method; keep the order in which requests were sent
-      # for method in "${methods[@]}" ; do
-      #  egrep '^\| *'"${browser}"' *\| *'"${method}"' *\| *'"${origin}"' *\| *'"${creds}"' *\|' "${TMPFILE}"
-      # done
-      egrep '^\| *'"${browser}"' *\| *'"[^\|]+"' *\| *'"${origin}"' *\| *'"${creds}"' *\|' "${TMPFILE}"
+      for method in "${exfMethods[@]}" ; do
+        [[ $DEBUG -eq 0 ]] || echo '^\| *'"${browser}"' *\| *'"[A-Z]+ \(${method}\)"' *\| *'"${origin}"' *\| *'"${creds}"' *\|' 1>&2
+        egrep '^\| *'"${browser}"' *\| *'"[A-Z]+ \(${method}\)"' *\| *'"${origin}"' *\| *'"${creds}"' *\|' "${TMPFILE}"
+      done
       echo "${sep}"
     done
   done
 done | uniq >> "${OUTFILE}"
 
-rm "${TMPFILE}"
+[[ $DEBUG -eq 0 ]] && rm "${TMPFILE}"
